paypal modified

+++ app/Http/Controllers/Instructor/CourseController.php
@@ -152,7 +152,10 @@ class CourseController extends Controller
     {
         $course->status = 2;
         $course->save();
-        $course->observation->delete();
+        if ($course->observation) {
+            # code...
+            $course->observation->delete();
+        }
         return redirect()->route('instructor.courses.edit', $course);
         // return 'Aquí se cambiará el estado del curso';
     }
